pyproject.toml
[project]
name = "ai-persona-system"
version = "0.2.2"
description = "AIペルソナシステム - インタビューからペルソナを生成し、議論を通じてインサイトを生成"
authors = [
    {name = "AI Persona Team"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "python-dotenv>=1.0.0",
    "strands-agents>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "bedrock-agentcore>=0.0.6",
    "strands-agents-tools>=0.2.19",
    "markitdown[pdf,docx,pptx,xlsx]>=0.1.4",
    "cachetools>=5.3.0",
    "datasets>=3.0.0",
    "duckdb>=1.1.0",
    "polars>=1.0.0",
    "markdown>=3.5.0",
    "nh3>=0.2.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.5.0",
    "moto[dynamodb,s3]>=5.0.0",
    "hypothesis>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
ai-persona = "src.app:main"
